Ajusta e move função geraHashCSRT
